reference/pyproject.toml
[project]
name = "takopi-discord"
authors = [{name = "ayvee"}]
version = "0.3.0"
description = "Discord transport plugin for takopi"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.14"
dependencies = [
    "py-cord>=2.6.0",
    "openai>=1.0.0",
    "takopi>=0.17.0",
    "pywhispercpp>=1.4.0",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/asianviking/takopi-discord"
Repository = "https://github.com/asianviking/takopi-discord"
Issues = "https://github.com/asianviking/takopi-discord/issues"

[project.entry-points."takopi.transport_backends"]
discord = "takopi_discord.backend:BACKEND"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/takopi_discord"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-anyio>=0.0.0",
    "ruff>=0.14.0",
]

[tool.ruff.lint]
extend-select = ["B", "BLE001", "C4", "PERF", "SIM", "UP"]
